fix: Fix getting user and userobj from g object
With the Pylons g object, if we tried to access a missing attribute (e.g. g.user), we were just sent on our way without an error.

With the Flask g object, we have to check that g has the attribute first; if we try to access it and it's not there, we get an error. But even if the g object has the attribute, its value might still be None, causing an error when we try to do something with it. This means we need both checks.